Find more anonymous strings
If we find paths in a json file (e.g. .babelrc), we can expand it into a
list of paths and assume that they are all used. For instance,

  src/main.js

in a packages/foo/package.json file is likely to mean either

./src/main.js (assuming process.cwd as root)

or

./packages/foo/src/main.js (assuming relative to file)

We can afford to check both of these paths.